refactor(ui): promote V4 run-detail components to canonical RunDetails/shared
Actions, RunInfo, TopInfo, StepInfo, UserlandAttrs, MetadataAttrs,
runDetailsUtils→utils, and types all move from RunDetailsV4/ into the
shared/ folder. V4 variants are strict additive supersets of V3 (new
optional props like isDurableEndpoint, richer rendering, additional
tabs), so V3 callers inherit them safely:

- isDurableEndpoint defaults to false in Actions / RunInfo / StepInfo,
  preserving V3's Rerun-button behavior.
- V4-only tabs in StepInfo/TopInfo (response headers, experiment metadata,
  cron trigger metadata) gate on presence of their data, so V3 users
  don't see extra tabs.
- V3 users now receive V4's richer UserlandAttrs (resource attrs, span
  metadata) and MetadataAttrs (semantic groups), plus V4's
  formatMilliseconds format ("1m 30s" rather than "1.5m").
- V3's `enable-step-metadata` feature flag is dropped — V4 renders
  metadata unconditionally when present.

External consumers updated to the new types path: graphql.config.ts,
gql/graphql.ts, Debugger/*, SharedContext/useGetDebugRun, useGetDebugSession,
useGetRun, useGetRunTrace, Table/Cell.

Both RunDetailsV3/ and RunDetailsV4/ shells remain; the run-details-v4
feature flag at the dashboard level still selects between them, but they
now share the same internals.
