+++ ui/packages/components/src/Debugger/DebugTrace.tsx
@@ -4,8 +4,8 @@ import { RiArrowRightSLine } from '@remixicon/react';
 import { InlineSpans } from '../RunDetailsV3/InlineSpans';
 import { StepType } from '../RunDetailsV3/StepType';
 import { TimelineHeader } from '../RunDetailsV3/TimelineHeader';
-import { type Trace } from '../RunDetailsV3/types';
-import { getSpanName, traceHasChildren, useStepSelection } from '../RunDetailsV3/utils';
+import { type Trace } from '../RunDetails/shared/types';
+import { getSpanName, traceHasChildren, useStepSelection } from '../RunDetails/shared/utils';
 import { overlayDebugRuns } from './utils';
 
 type Props = {
