+++ ui/packages/components/src/RunDetails/shared/StepInfo.tsx
@@ -1,41 +1,54 @@
 import { useEffect, useState } from 'react';
-import { Button as NewButton } from '@inngest/components/Button';
-import { Button } from '@inngest/components/Button/Button';
 import { RiArrowRightSLine } from '@remixicon/react';
 
-import { AITrace } from '../AI/AITrace';
-import { parseAIOutput } from '../AI/utils';
+import { AITrace } from '../../AI/AITrace';
+import { parseAIOutput } from '../../AI/utils';
+import { Button } from '../../Button/Button';
 import {
   CodeElement,
   ElementWrapper,
+  IDElement,
   LinkElement,
   TextElement,
   TimeElement,
-} from '../DetailsCard/Element';
-import { RerunModal as NewRerunModal, RerunModal } from '../Rerun/RerunModal';
-import { useShared } from '../SharedContext/SharedContext';
-import { useBooleanFlag } from '../SharedContext/useBooleanFlag';
-import { useGetTraceResult } from '../SharedContext/useGetTraceResult';
-import { usePathCreator } from '../SharedContext/usePathCreator';
-import { Time } from '../Time';
-import { usePrettyErrorBody, usePrettyJson, usePrettyShortError } from '../hooks/usePrettyJson';
-import { formatMilliseconds, toMaybeDate } from '../utils/date';
-import { ErrorInfo } from '../RunDetails/shared/ErrorInfo';
-import { IO } from '../RunDetails/shared/IO';
+} from '../../DetailsCard/Element';
+import { Pill } from '../../Pill/Pill';
+import { RerunModal } from '../../Rerun/RerunModal';
+import { useShared } from '../../SharedContext/SharedContext';
+import { useGetTraceResult } from '../../SharedContext/useGetTraceResult';
+import { usePathCreator } from '../../SharedContext/usePathCreator';
+import { getStatusBackgroundClass, getStatusTextClass } from '../../Status/statusClasses';
+import { Time } from '../../Time';
+import type { SpanMetadataKind } from '../../generated';
+import { usePrettyErrorBody, usePrettyJson, usePrettyShortError } from '../../hooks/usePrettyJson';
+import { toMaybeDate } from '../../utils/date';
+import { ErrorInfo } from './ErrorInfo';
+import { IO } from './IO';
 import { MetadataAttrs } from './MetadataAttrs';
-import { Tabs } from '../RunDetails/shared/Tabs';
+import { Tabs } from './Tabs';
 import { UserlandAttrs } from './UserlandAttrs';
+import { formatDuration, maybeBooleanToString, type StepInfoType } from './utils';
 import {
+  isExperimentMetadata,
   isStepInfoInvoke,
   isStepInfoSignal,
   isStepInfoSleep,
   isStepInfoWait,
+  type SpanMetadataScope,
   type StepInfoInvoke,
   type StepInfoSignal,
   type StepInfoSleep,
   type StepInfoWait,
 } from './types';
-import { maybeBooleanToString, type StepInfoType } from './utils';
+
+const STEP_OP_LABELS: Record<string, string> = {
+  RUN: 'step.run',
+  INVOKE: 'step.invoke',
+  SLEEP: 'step.sleep',
+  WAIT_FOR_EVENT: 'step.waitForEvent',
+  AI_GATEWAY: 'step.ai',
+  WAIT_FOR_SIGNAL: 'step.waitForSignal',
+};
 
 type StepKindInfoProps = {
   stepInfo: StepInfoType['trace']['stepInfo'];
@@ -46,7 +59,17 @@ const InvokeInfo = ({ stepInfo }: { stepInfo: StepInfoInvoke }) => {
   const timeout = toMaybeDate(stepInfo.timeout);
   return (
     <>
-      <ElementWrapper label="Run">
+      <ElementWrapper label="Function">
+        <LinkElement href={pathCreator.function({ functionSlug: stepInfo.functionID })}>
+          {stepInfo.functionID}
+        </LinkElement>
+      </ElementWrapper>
+      <ElementWrapper label="Triggering Event ID">
+        <LinkElement href={pathCreator.eventPopout({ eventID: stepInfo.triggeringEventID })}>
+          {stepInfo.triggeringEventID}
+        </LinkElement>
+      </ElementWrapper>
+      <ElementWrapper label="Triggered Run ID">
         {stepInfo.runID ? (
           <LinkElement href={pathCreator.runPopout({ runID: stepInfo.runID })}>
             {stepInfo.runID}
@@ -61,6 +84,13 @@ const InvokeInfo = ({ stepInfo }: { stepInfo: StepInfoInvoke }) => {
       <ElementWrapper label="Timed out">
         <TextElement>{maybeBooleanToString(stepInfo.timedOut)}</TextElement>
       </ElementWrapper>
+      {stepInfo.returnEventID && (
+        <ElementWrapper label="Return Event ID">
+          <LinkElement href={pathCreator.eventPopout({ eventID: stepInfo.returnEventID })}>
+            {stepInfo.returnEventID}
+          </LinkElement>
+        </ElementWrapper>
+      )}
     </>
   );
 };
@@ -75,6 +105,7 @@ const SleepInfo = ({ stepInfo }: { stepInfo: StepInfoSleep }) => {
 };
 
 const WaitInfo = ({ stepInfo }: { stepInfo: StepInfoWait }) => {
+  const { pathCreator } = usePathCreator();
   const timeout = toMaybeDate(stepInfo.timeout);
   return (
     <>
@@ -87,6 +118,13 @@ const WaitInfo = ({ stepInfo }: { stepInfo: StepInfoWait }) => {
       <ElementWrapper label="Timed out">
         <TextElement>{maybeBooleanToString(stepInfo.timedOut)}</TextElement>
       </ElementWrapper>
+      {stepInfo.foundEventID && (
+        <ElementWrapper label="Matched Event ID">
+          <LinkElement href={pathCreator.eventPopout({ eventID: stepInfo.foundEventID })}>
+            {stepInfo.foundEventID}
+          </LinkElement>
+        </ElementWrapper>
+      )}
       <ElementWrapper className="w-full" label="Match expression">
         {stepInfo.expression ? (
           <CodeElement value={stepInfo.expression} />
@@ -131,13 +169,13 @@ export const StepInfo = ({
   pollInterval: initialPollInterval,
   tracesPreviewEnabled,
   debug = false,
-  newStack = false,
+  isDurableEndpoint,
 }: {
   selectedStep: StepInfoType;
   pollInterval?: number;
   tracesPreviewEnabled?: boolean;
   debug?: boolean;
-  newStack?: boolean;
+  isDurableEndpoint?: boolean;
 }) => {
   const { cloud } = useShared();
   const [expanded, setExpanded] = useState(true);
@@ -150,14 +188,13 @@ export const StepInfo = ({
     preview: tracesPreviewEnabled,
   });
 
-  const { booleanFlag } = useBooleanFlag();
-  const { value: metadataIsEnabled } = booleanFlag('enable-step-metadata', false);
+  const metadataIsEnabled = true;
 
   useEffect(() => {
     result && setPollInterval(undefined);
   }, [result]);
 
-  const delayText = formatMilliseconds(
+  const delayText = formatDuration(
     (toMaybeDate(trace.startedAt) ?? new Date()).getTime() - new Date(trace.queuedAt).getTime()
   );
 
@@ -165,7 +202,7 @@ export const StepInfo = ({
   const endedAt = toMaybeDate(trace.endedAt);
   const duration = startedAt ? (endedAt ?? new Date()).getTime() - startedAt.getTime() : 0;
 
-  const durationText = duration > 0 ? formatMilliseconds(duration) : '-';
+  const durationText = duration > 0 ? formatDuration(duration) : '-';
 
   const stepKindInfo = getStepKindInfo({
     stepInfo: trace.stepInfo,
@@ -176,6 +213,47 @@ export const StepInfo = ({
   const prettyOutput = usePrettyJson(result?.data ?? '') || (result?.data ?? '');
   const prettyErrorBody = usePrettyErrorBody(result?.error);
   const prettyShortError = usePrettyShortError(result?.error);
+  const showRerunFromStep =
+    !isDurableEndpoint && !debug && runID && trace.stepID && (!cloud || prettyInput);
+
+  const responseHeaderMetadata = trace.metadata?.filter(
+    (md) => md.kind === 'inngest.response_headers'
+  );
+
+  // TODO: remove metadata handling once all response header
+  // data in history uses the response field. (After 2026-06-03)
+  const responseHeaderData = responseHeaderMetadata?.length
+    ? responseHeaderMetadata
+    : trace.response
+    ? [
+        {
+          kind: 'inngest.response_headers' as SpanMetadataKind,
+          values: {
+            ...Object.fromEntries(
+              Object.entries(trace.response?.headers ?? {}).map(([k, v]) => [
+                k,
+                Array.isArray(v) ? v.join(', ') : v,
+              ])
+            ),
+            'Status Code': trace.response.statusCode.toString(),
+          },
+          updatedAt: trace.endedAt ?? trace.startedAt ?? trace.queuedAt,
+          scope: 'step_attempt' as SpanMetadataScope,
+        },
+      ]
+    : [];
+
+  const nonHeaderMetadata = metadataIsEnabled
+    ? trace.metadata?.filter(
+        (md) => md.kind !== 'inngest.response_headers' && !isExperimentMetadata(md)
+      )
+    : undefined;
+
+  const experimentMetadataList = metadataIsEnabled
+    ? trace.metadata?.filter(isExperimentMetadata)
+    : undefined;
+
+  const experimentMetadata = experimentMetadataList?.[0];
 
   const hasNoData = !loading && !prettyInput && !prettyOutput && !result?.error;
 
@@ -200,46 +278,36 @@ export const StepInfo = ({
           />
 
           <span className="text-basis text-sm font-normal">{trace.name}</span>
+          {trace.attempts !== null && trace.attempts > 0 && (
+            <span data-testid="retry-attempt-badge">
+              <Pill
+                className={`${getStatusBackgroundClass(trace.status)} ${getStatusTextClass(
+                  trace.status
+                )}`}
+              >
+                {trace.attempts} {trace.attempts === 1 ? 'retry' : 'retries'}
+              </Pill>
+            </span>
+          )}
         </div>
-        {!debug &&
-          runID &&
-          trace.stepID &&
-          (!cloud || prettyInput) &&
-          (newStack ? (
-            <>
-              <NewButton
-                kind="primary"
-                appearance="outlined"
-                size="medium"
-                label="Rerun from step"
-                onClick={() => setRerunModalOpen(true)}
-              />
-              <NewRerunModal
-                open={rerunModalOpen}
-                setOpen={setRerunModalOpen}
-                runID={runID}
-                stepID={trace.stepID}
-                input={prettyInput || result?.input || ''}
-              />
-            </>
-          ) : (
-            <>
-              <Button
-                kind="primary"
-                appearance="outlined"
-                size="medium"
-                label="Rerun from step"
-                onClick={() => setRerunModalOpen(true)}
-              />
-              <RerunModal
-                open={rerunModalOpen}
-                setOpen={setRerunModalOpen}
-                runID={runID}
-                stepID={trace.stepID}
-                input={prettyInput || result?.input || ''}
-              />
-            </>
-          ))}
+        {showRerunFromStep && (
+          <>
+            <Button
+              kind="primary"
+              appearance="outlined"
+              size="medium"
+              label="Rerun from step"
+              onClick={() => setRerunModalOpen(true)}
+            />
+            <RerunModal
+              open={rerunModalOpen}
+              setOpen={setRerunModalOpen}
+              runID={runID}
+              stepID={trace.stepID!}
+              input={prettyInput || result?.input || ''}
+            />
+          </>
+        )}
       </div>
 
       {expanded && (
@@ -276,7 +344,30 @@ export const StepInfo = ({
             <TextElement>{durationText}</TextElement>
           </ElementWrapper>
 
+          {trace.stepOp && (
+            <ElementWrapper label="Step Type">
+              <CodeElement value={STEP_OP_LABELS[trace.stepOp] ?? trace.stepOp} />
+            </ElementWrapper>
+          )}
+
           {stepKindInfo}
+
+          {experimentMetadata && (
+            <>
+              <ElementWrapper label="Experiment name">
+                <TextElement>{experimentMetadata.values.experiment_name}</TextElement>
+              </ElementWrapper>
+              <ElementWrapper label="Variant">
+                <TextElement>{experimentMetadata.values.variant_selected}</TextElement>
+              </ElementWrapper>
+            </>
+          )}
+
+          {debug && trace.debugRunID && (
+            <ElementWrapper label="Debug Run ID">
+              <IDElement>{trace.debugRunID}</IDElement>
+            </ElementWrapper>
+          )}
 
           {aiOutput && <AITrace aiOutput={aiOutput} />}
         </div>
@@ -292,12 +383,30 @@ export const StepInfo = ({
                 id: 'attributes',
                 node: <UserlandAttrs userlandSpan={trace.userlandSpan} />,
               },
-              ...(metadataIsEnabled && trace.metadata?.length
+              ...(responseHeaderData?.length
+                ? [
+                    {
+                      label: 'Headers',
+                      id: 'headers',
+                      node: <MetadataAttrs metadata={responseHeaderData} />,
+                    },
+                  ]
+                : []),
+              ...(experimentMetadataList?.length
+                ? [
+                    {
+                      label: 'Experiment',
+                      id: 'experiment',
+                      node: <MetadataAttrs metadata={experimentMetadataList} />,
+                    },
+                  ]
+                : []),
+              ...(nonHeaderMetadata?.length
                 ? [
                     {
                       label: 'Metadata',
                       id: 'metadata',
-                      node: <MetadataAttrs metadata={trace.metadata} />,
+                      node: <MetadataAttrs metadata={nonHeaderMetadata} />,
                     },
                   ]
                 : []),
@@ -350,12 +459,30 @@ export const StepInfo = ({
                         },
                       ]
                     : []),
-                  ...(metadataIsEnabled && trace.metadata?.length
+                  ...(responseHeaderData?.length
+                    ? [
+                        {
+                          label: 'Headers',
+                          id: 'headers',
+                          node: <MetadataAttrs metadata={responseHeaderData} />,
+                        },
+                      ]
+                    : []),
+                  ...(experimentMetadataList?.length
+                    ? [
+                        {
+                          label: 'Experiment',
+                          id: 'experiment',
+                          node: <MetadataAttrs metadata={experimentMetadataList} />,
+                        },
+                      ]
+                    : []),
+                  ...(nonHeaderMetadata?.length
                     ? [
                         {
                           label: 'Metadata',
                           id: 'metadata',
-                          node: <MetadataAttrs metadata={trace.metadata} />,
+                          node: <MetadataAttrs metadata={nonHeaderMetadata} />,
                         },
                       ]
                     : []),
